pyproject.toml
[project]
name = "arc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arc-agi>=0.9.1",
    "dspy>=3.1.3",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
]
